(#86) fix: WorkSession の Sidekiq フォールバック分割と保護を追加
- Sidekiq/ジョブ関連の private メソッドを Concern に分割
- Sidekiq 未導入時の require 保護、push 失敗時の ActiveJob フォールバックを実装
- RuboCop の Metrics/ClassLength を解消
